+++ src/front/glsl/mod.rs
@@ -10,7 +10,7 @@ pub use ast::{Precision, Profile};
 pub use error::{Error, ErrorKind, ExpectedToken};
 pub use token::TokenValue;
 
-use crate::{FastHashMap, FastHashSet, Handle, Module, ShaderStage, Span, Type};
+use crate::{proc::Layouter, FastHashMap, FastHashSet, Handle, Module, ShaderStage, Span, Type};
 use ast::{EntryArg, FunctionDeclaration, GlobalLookup};
 use parser::ParsingContext;
 
@@ -166,6 +166,8 @@ pub struct Parser {
 
     entry_args: Vec<EntryArg>,
 
+    layouter: Layouter,
+
     errors: Vec<Error>,
 
     module: Module,
@@ -179,6 +181,7 @@ impl Parser {
         self.lookup_type.clear();
         self.global_variables.clear();
         self.entry_args.clear();
+        self.layouter.clear();
 
         // This is necessary because if the last parsing errored out, the module
         // wouldn't have been swapped