+++ src/front/wgsl/tests.rs
@@ -398,7 +398,7 @@ fn parse_array_length() {
     parse_str(
         "
         struct Foo {
-            data: @stride(4) array<u32>;
+            data: array<u32>;
         }; // this is used as both input and output for convenience
 
         @group(0) @binding(0)