+++ tests/out/glsl/access.atomics.Compute.glsl
@@ -5,12 +5,15 @@ precision highp int;
 
 layout(local_size_x = 1, local_size_y = 1, local_size_z = 1) in;
 
+struct AlignedWrapper {
+    int value;
+};
 layout(std430) buffer Bar_block_0Compute {
     mat4x4 matrix;
     mat2x2 matrix_array[2];
     int atom;
     uvec2 arr[2];
-    int data[];
+    AlignedWrapper data[];
 } _group_0_binding_0_cs;